+++ src/data-workspace/data-details-sidebar/comment.js
@@ -4,8 +4,7 @@ import PropTypes from 'prop-types'
 import React, { useState } from 'react'
 import { Form, Field } from 'react-final-form'
 import { ExpandableUnit } from '../../shared/index.js'
-import { useSetDataValueMutation } from '../use-data-value-mutation/_data-value-mutations.js'
-// import { useSetDataValueCommentMutation } from '../use-data-value-mutation/index.js'
+import { useSetDataValueMutation } from '../use-data-value-mutation/index.js'
 import styles from './comment.module.css'
 import LoadingError from './loading-error.js'
