+++ src/data-workspace/inputs/boolean-radios.js
@@ -3,8 +3,7 @@ import { Button, Radio } from '@dhis2/ui'
 import cx from 'classnames'
 import React, { useState } from 'react'
 import { useField, useForm } from 'react-final-form'
-import { useSetDataValueMutation } from '../use-data-value-mutation/_data-value-mutations.js'
-// import { useSetDataValueMutation } from '../use-data-value-mutation/index.js'
+import { useSetDataValueMutation } from '../use-data-value-mutation/index.js'
 import styles from './inputs.module.css'
 import { convertCallbackSignatures, InputPropTypes } from './utils.js'
