+++ src/data-workspace/inputs/file-inputs.js
@@ -7,8 +7,7 @@ import { useQuery } from 'react-query'
 import {
     useDeleteDataValueMutation,
     useUploadFileDataValueMutation,
-} from '../use-data-value-mutation/_data-value-mutations.js'
-// } from '../use-data-value-mutation/index.js'
+} from '../use-data-value-mutation/index.js'
 import useFileInputUrl from '../use-file-input-url.js'
 import styles from './inputs.module.css'
