--- src/data-workspace/use-data-value-mutation/get-data-value-index.js
@@ -1,12 +1,0 @@
-export default function getDataValueIndex(
-    previousDataValues,
-    { co, de, ou, pe }
-) {
-    return previousDataValues.findIndex(
-        (dataValue) =>
-            dataValue.categoryOptionCombo === co &&
-            dataValue.dataElement === de &&
-            dataValue.orgUnit === ou &&
-            dataValue.period === pe
-    )
-}
