+++ src/data-workspace/use-data-value-mutation/index.js
@@ -1,4 +1,6 @@
-export { useDeleteDataValueMutation } from './use-delete-data-value-mutation.js'
-export { useSetDataValueMutation } from './use-set-data-value-mutation.js'
-export { useSetDataValueCommentMutation } from './use-set-data-value-comment-mutation.js'
-export { useUploadDataValueFileMutation } from './use-upload-data-value-file-mutation.js'
+export {
+    useSetDataValueMutation,
+    useDeleteDataValueMutation,
+    useUploadFileDataValueMutation,
+} from './data-value-mutations.js'
+export { useDataValueParams } from './use-data-value-params.js'
